fix: adjusting the prospect service by code
