+++ src/services/prospects/ProspectByCode/index.ts
@@ -5,10 +5,10 @@ import {
 } from "@config/environment";
 import { IProspect } from "./types";
 
-const getAllProspects = async (
+const getSearchProspectByCode = async (
   businessUnitPublicCode: string,
   prospectCode: string
-): Promise<IProspect | IProspect[]> => {
+): Promise<IProspect> => {
   const maxRetries = maxRetriesServices;
   const fetchTimeout = fetchTimeoutServices;
 
@@ -52,7 +52,7 @@ const getAllProspects = async (
       }
 
       if (Array.isArray(data)) {
-        return data as IProspect[];
+        return data[0] as IProspect;
       }
 
       return data;
@@ -69,4 +69,4 @@ const getAllProspects = async (
   throw new Error("No se pudo obtener la tarea después de varios intentos.");
 };
 
-export { getAllProspects };
+export { getSearchProspectByCode };
